Implemented A-Packages for finnbot
