+++ src/main/java/finnbot/command/Command.java
@@ -1,3 +1,8 @@
+package finnbot.command;
+import finnbot.Storage;
+import finnbot.TasksList;
+import finnbot.Ui;
+
 public abstract class Command {
     public abstract void execute(TasksList tasks, Ui ui, Storage storage);
