+++ src/main/java/finnbot/command/DeleteCommand.java
@@ -1,15 +1,20 @@
+package finnbot.command;
+import finnbot.Storage;
+import finnbot.TasksList;
+import finnbot.Ui;
+
 import java.io.File;
 
-public class MarkCommand extends Command {
+public class DeleteCommand extends Command {
+
     private final int taskIndex;
 
-    public MarkCommand(int taskIndex) {
+    public DeleteCommand(int taskIndex) {
         this.taskIndex = taskIndex;
     }
 
     public void execute(TasksList tasks, Ui ui, Storage storage) {
-        tasks.taskMarker(taskIndex, true);
+        tasks.deleteTask(taskIndex);
         storage.saveFile(tasks, new File(Storage.FILEPATH));
     }
-
 }
