+++ src/main/java/finnbot/command/MarkCommand.java
@@ -1,14 +1,20 @@
+package finnbot.command;
+import finnbot.Storage;
+import finnbot.TasksList;
+import finnbot.Ui;
+
 import java.io.File;
 
-public class UnmarkCommand extends Command {
+public class MarkCommand extends Command {
     private final int taskIndex;
 
-    public UnmarkCommand(int taskIndex) {
+    public MarkCommand(int taskIndex) {
         this.taskIndex = taskIndex;
     }
 
     public void execute(TasksList tasks, Ui ui, Storage storage) {
-        tasks.taskMarker(taskIndex, false);
+        tasks.taskMarker(taskIndex, true);
         storage.saveFile(tasks, new File(Storage.FILEPATH));
     }
+
 }
