+++ src/main/java/finnbot/command/UnmarkCommand.java
@@ -1,0 +1,19 @@
+package finnbot.command;
+import finnbot.Storage;
+import finnbot.TasksList;
+import finnbot.Ui;
+
+import java.io.File;
+
+public class UnmarkCommand extends Command {
+    private final int taskIndex;
+
+    public UnmarkCommand(int taskIndex) {
+        this.taskIndex = taskIndex;
+    }
+
+    public void execute(TasksList tasks, Ui ui, Storage storage) {
+        tasks.taskMarker(taskIndex, false);
+        storage.saveFile(tasks, new File(Storage.FILEPATH));
+    }
+}
