+++ src/main/java/finnbot/exceptions/EmptyInputException.java
@@ -1,3 +1,5 @@
+package finnbot.exceptions;
+
 public class EmptyInputException extends Exception {
     public EmptyInputException(String message) {
         super(message);
