+++ src/main/java/finnbot/exceptions/InvalidCommandException.java
@@ -1,3 +1,5 @@
+package finnbot.exceptions;
+
 public class InvalidCommandException extends Exception {
     public InvalidCommandException(String message) {
         super(message);
