+++ src/main/java/finnbot/tasks/Tasks.java
@@ -1,6 +1,8 @@
+package finnbot.tasks;
+
 public class Tasks {
     protected String description;
-    protected boolean isDone;
+    public boolean isDone;
 
     public Tasks(String description) {
         this.description = description;
