+++ src/main/java/finnbot/tasks/ToDos.java
@@ -1,5 +1,7 @@
+package finnbot.tasks;
+
 public class ToDos extends Tasks {
-    //constructor for ToDos class
+
     public ToDos(String description) {
         super(description);
     }
